Rename file to match class name #2714
