+++ data-serving/reusable-data-service/reusable_data_service/model/case.py
@@ -10,7 +10,7 @@ from reusable_data_service.util.errors import (
     PreconditionUnsatisfiedError,
     ValidationError,
 )
-from reusable_data_service.util.iso_json_encoder import DataServiceJSONEncoder
+from reusable_data_service.util.data_service_json_encoder import DataServiceJSONEncoder
 
 
 @dataclasses.dataclass()
